+++ android-apps/clipnotes/app/src/main/java/com/yasirarafat/clipnotes/data/AppDao.kt
@@ -1,4 +1,4 @@
-package com.shishurmedhabikash.clipnotes.data
+package com.yasirarafat.clipnotes.data
 
 import androidx.room.Dao
 import androidx.room.Insert