+++ android-apps/clipnotes/app/src/main/java/com/yasirarafat/clipnotes/ui/screens/CategoriesScreen.kt
@@ -1,4 +1,4 @@
-package com.shishurmedhabikash.clipnotes.ui.screens
+package com.yasirarafat.clipnotes.ui.screens
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -38,7 +38,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.shishurmedhabikash.clipnotes.data.Category
+import com.yasirarafat.clipnotes.data.Category
 
 @Composable
 fun CategoriesScreen(