+++ android-apps/clipnotes/app/src/main/java/com/yasirarafat/clipnotes/ui/screens/SettingsScreen.kt
@@ -1,4 +1,4 @@
-package com.shishurmedhabikash.clipnotes.ui.screens
+package com.yasirarafat.clipnotes.ui.screens
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
@@ -19,7 +19,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.shishurmedhabikash.clipnotes.ui.NotesViewModel
+import com.yasirarafat.clipnotes.ui.NotesViewModel
 
 @Composable
 fun SettingsScreen(vm: NotesViewModel) {