+++ android-apps/clipnotes/app/src/main/java/com/yasirarafat/clipnotes/ui/screens/TrashScreen.kt
@@ -1,4 +1,4 @@
-package com.shishurmedhabikash.clipnotes.ui.screens
+package com.yasirarafat.clipnotes.ui.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -32,8 +32,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.shishurmedhabikash.clipnotes.data.Category
-import com.shishurmedhabikash.clipnotes.data.Note
+import com.yasirarafat.clipnotes.data.Category
+import com.yasirarafat.clipnotes.data.Note
 
 @Composable
 fun TrashScreen(