+++ android-apps/clipnotes/app/src/main/java/com/yasirarafat/clipnotes/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.shishurmedhabikash.clipnotes.ui.theme
+package com.yasirarafat.clipnotes.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme